+++ .cspell.json
@@ -31,6 +31,7 @@
     "Plutomi",
     "praguru",
     "Sider",
+    "tabler",
     "tailwindcss",
     "TLDR",
     "trenchard",
