+++ packages/validation/index.ts
@@ -1,3 +1,5 @@
 import * as Schema from "./schemas";
 
+export * from "./validate";
+
 export { Schema };
