+++ packages/web/pages/_app.tsx
@@ -7,7 +7,13 @@ import { Inter } from "next/font/google";
 const inter = Inter({ subsets: ["latin"] });
 
 const App = ({ Component, pageProps }: AppProps) => (
-  <MantineProvider withGlobalStyles withNormalizeCSS>
+  <MantineProvider
+    withGlobalStyles
+    withNormalizeCSS
+    theme={{
+      primaryColor: "indigo"
+    }}
+  >
     <Notifications position="top-center" />
 
     <Component {...pageProps} className={inter.className} />
