+++ packages/web/pages/index.tsx
@@ -1,14 +1,12 @@
 import LandingHero from "@/components/LandingHero/LandingHero";
 import { WaitListCard } from "@/components/WaitistCard";
-import { NextPage } from "next";
+import type { NextPage } from "next";
 
-const Home: NextPage = () => {
-  return (
-    <>
-      <LandingHero />
-      <WaitListCard />
-    </>
-  );
-};
+const Home: NextPage = () => (
+  <>
+    <LandingHero />
+    <WaitListCard />
+  </>
+);
 
 export default Home;
