+++ packages/web/utils/handleAxiosResponse.ts
@@ -1,0 +1,24 @@
+import axios, { type AxiosError } from "axios";
+
+type ErrorResponse = {
+  message: string;
+};
+
+export const handleAxiosError = <T extends ErrorResponse>(
+  err: unknown
+): string => {
+  if (axios.isAxiosError(err)) {
+    const axiosError = err as AxiosError<T>;
+    const { message } = axiosError.response?.data ?? {};
+    if (typeof message === "string") {
+      return message;
+    }
+    return "No additional information was provided by the server.";
+  }
+
+  if (err instanceof Error) {
+    return err.message;
+  }
+
+  return "An unknown error occurred.";
+};
